+++ apps/client/src/app/app-routing.module.ts
@@ -130,9 +130,7 @@ const routes: Routes = [
   {
     path: publicRoutes.start.path,
     loadChildren: () =>
-      import('./pages/landing/landing-page.module').then(
-        (m) => m.LandingPageModule
-      )
+      import('./pages/landing/landing-page.routes').then((m) => m.routes)
   },
   {
     loadComponent: () =>