+++ libs/common/src/lib/interfaces/responses/watchlist-response.interface.ts
@@ -8,5 +8,7 @@ export interface WatchlistResponse {
     marketCondition: Benchmark['marketCondition'];
     name: string;
     performances: Benchmark['performances'];
+    trend50d: Benchmark['trend50d'];
+    trend200d: Benchmark['trend200d'];
   })[];
 }