+++ libs/common/src/lib/interfaces/asset-class-selector-option.interface.ts
@@ -1,0 +1,6 @@
+import { AssetClass, AssetSubClass } from '@prisma/client';
+
+export interface AssetClassSelectorOption {
+  id: AssetClass | AssetSubClass;
+  label: string;
+}