+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored the register page to standalone
 - Migrated the login with access token dialog from `ngModel` to form control
+- Upgraded `chart.js` from version `4.4.9` to `4.5.0`
 
 ## 2.193.0 - 2025-08-22