+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored the register page to standalone
 - Migrated the login with access token dialog from `ngModel` to form control
+- Upgraded `ionicons` from version `8.0.10` to `8.0.13`
 - Upgraded `prisma` from version `6.12.0` to `6.14.0`
 
 ## 2.193.0 - 2025-08-22