+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactored the dialog footer component to standalone
 - Refactored the dialog header component to standalone
+- Refactored the landing page to standalone
 - Refactored the register page to standalone
 - Migrated the login with access token dialog from `ngModel` to form control
 - Upgraded `prisma` from version `6.12.0` to `6.14.0`