+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored the dialog footer component to standalone
 - Refactored the dialog header component to standalone
 - Refactored the landing page to standalone
+- Refactored the pricing page to standalone
 - Refactored the register page to standalone
 - Migrated the login with access token dialog from `ngModel` to form control
 - Upgraded `@ionic/angular` from version `8.6.3` to `8.7.3`